Restructuring to modularise: - created dpv folder with separate helper and validator functions - added extra describe statements to section out tests - added step by step comments for longer functions - removed redundant code - amended cypress config to ignore dpv helpers and validators
Also tweaked submission time check for self assessment page completion tag to give double-digit minutes in all instances and moved function call to later in thread to reduce risk of off-by-one errors
