+++ tests/Dfe.PlanTech.Web.E2ETests/cypress/e2e/dynamic-page-validator/dpv-helpers/data-loaded-check.js
@@ -1,0 +1,16 @@
+/**
+ * Check if data has been loaded and log a message if not.
+ *
+ * @param {Array | Map} contentMap - the map of content
+ * @return {boolean} haveContent - whether data has been loaded
+ */
+
+export const dataLoaded = (contentMap) => {
+    const haveContent = contentMap && ((contentMap.size && contentMap.size > 0) || (contentMap.length && contentMap.length > 0));
+
+    if (!haveContent) {
+        console.log("Data has not been loaded");
+    }
+
+    return haveContent;
+}
