+++ tests/Dfe.PlanTech.Web.E2ETests/cypress/e2e/dynamic-page-validator/dpv-validators/answer-validator.js
@@ -1,0 +1,7 @@
+export const validateAnswers = (matchingQuestion) => {
+    for (const answer of matchingQuestion.answers) {
+        cy.get(
+            "div.govuk-radios div.govuk-radios__item label.govuk-radios__label.govuk-label"
+        ).contains(answer.text);
+    }
+}
