+++ tests/Dfe.PlanTech.Web.E2ETests/cypress/e2e/dynamic-page-validator/dpv-validators/nav-links-validator.js
@@ -1,0 +1,18 @@
+import ValidateContent from "../../../helpers/content-validators/content-validator.js";
+import { dataLoaded } from "../dpv-helpers/data-loaded-check.js";
+
+export const validateNavigationLinks = (dataMapper) => {
+    if (!dataMapper) {
+        return;
+    }
+
+    const navigationLinks = Array.from(dataMapper.contents.get("navigationLink").entries());
+
+    if (!dataLoaded(navigationLinks)) {
+        return;
+    }
+
+    for (const [_, navigationLink] of navigationLinks) {
+        ValidateContent(navigationLink);
+    }
+}
