+++ tests/Dfe.PlanTech.Web.E2ETests/cypress/e2e/dynamic-page-validator/dpv-validators/recommendations-validator.js
@@ -1,0 +1,12 @@
+import { selfAssessmentPageNewRecommendation } from "./completion-tags-validator.js";
+
+/**
+ * Validates a recommendation for a specific maturity level.
+ *
+ * @param {Object} section - the section containing the recommendations
+ * @param {string} maturity - the maturity level to validate
+ */
+
+export const validateRecommendationForMaturity = (section, maturity) => {
+    selfAssessmentPageNewRecommendation(section, maturity);
+}
