+++ tests/Dfe.PlanTech.Web.E2ETests/cypress/helpers/page-slugs.js
@@ -1,1 +1,10 @@
 export const selfAssessmentSlug = "/self-assessment";
+
+export const FindPageForSlug = ({ slug, dataMapper }) => {
+    for (const [id, page] of dataMapper.pages.entries()) {
+        if (page.fields.slug == slug) {
+            return page;
+        }
+    }
+    return null;
+}
